Reorganize DocDB command registration
