+++ src/docdb/commands/createDocDBCollection.ts
@@ -1,0 +1,15 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { IActionContext } from "@microsoft/vscode-azext-utils";
+import { DocDBDatabaseTreeItem } from "../tree/DocDBDatabaseTreeItem";
+import { pickDocDBAccount } from "./pickDocDBAccount";
+
+export async function createDocDBCollection(context: IActionContext, node?: DocDBDatabaseTreeItem): Promise<void> {
+    if (!node) {
+        node = await pickDocDBAccount<DocDBDatabaseTreeItem>(context, DocDBDatabaseTreeItem.contextValue);
+    }
+    await node.createChild(context);
+}
